requirements.txt
# Core RAG System Dependencies (Python 3.11 + Windows Compatible)
chromadb==0.4.22
sentence-transformers>=2.3.0
huggingface-hub>=0.20.0
transformers>=4.35.0
tokenizers>=0.15.0

# PyTorch for Python 3.11 + Windows (minimum version 2.0.0)
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# PDF Processing
PyMuPDF==1.23.8
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.0.1

# Data Processing
pandas==1.5.3
numpy==1.24.3

# LLM API
groq

# Environment & Config
python-dotenv==1.0.0

# Web Framework
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# Voice Processing (using APIs - no complex dependencies)
gtts==2.5.0
pydub==0.25.1

# Authentication & Security
flask-jwt-extended==4.6.0
flask-bcrypt==1.0.1
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9

# Caching & Sessions
redis==5.0.1
flask-session==0.5.0

# Utilities
tqdm==4.65.0
opencv-python==4.8.1.78
matplotlib==3.8.2

# Compatible protobuf version
protobuf==4.25.3
